Add call to finish_up() to add final brace pair
Have been getting an error message for the very last variable written
out during a full documentation pass because the final pair of
braces has not been closed out.  Here's the message I get:

Warning: @section [gen_bdprob.R#10]: mismatched braces or quote

The problem was that I only called finish_up() if it hit an EOF,
but usually this condition does not occur because there's more
output in the log file after the end of the last echoed variable.
This meant that the loop ended correctly but never hit EOF so
never triggered the closing braces.